rename client to proxy
