+++ cmd/bootcamp/main.go
@@ -2,13 +2,13 @@ package main
 
 import (
 	"fmt"
-	"grabvn-golang-bootcamp/internal/bootcamp/feedbackclient"
+	"grabvn-golang-bootcamp/internal/bootcamp/feedbackproxy"
 	"grabvn-golang-bootcamp/internal/bootcamp/feedbackserver"
 )
 
 func main() {
 	go feedbackserver.StartServer()
-	go feedbackclient.StartClient()
+	go feedbackproxy.StartProxy()
 
 	_, _ = fmt.Scanln()
 }
