--- internal/bootcamp/feedbackclient/endpoint.go
@@ -1,1 +1,0 @@
-package feedbackclient
