+++ internal/bootcamp/feedbackproxy/endpoint.go
@@ -1,0 +1,1 @@
+package feedbackproxy
